bt: separate app_ble to hci and gap
